Add back missing newline in github-deploy-key (#1794)
In #1791 a `strip()` was added to a bunch of strings read from files. However, this also removes newlines, so for the github-deploy-key the `known_hosts` would end up on the same line as the end of the private key and therefore not work in the `source-controller`.

Instead of removing the `strip()` let's remove the `-` sign in the jinja end block so we keep newline after it.